requirements.txt
torch
opencv-python
numpy
ultralytics
pillow
transformers
torchvision
nsfw-detector